+++ CHANGELOG.md
@@ -5,6 +5,10 @@ Change Log
 ## UNRELEASED
 ##
 
+### Features
+* Adiciona endpoint de servico de pneus na API (PL-3645)
+* Cria DTO para a listagem de veículos na API (PL-3672)
+
 
 <a name="v3.4.0"></a>
 ## Version [v3.4.0](https://github.com/luizfp/PrologWebService/compare/v3.3.7...v3.4.0) (2021-06-07)
@@ -14,6 +18,7 @@ Change Log
 * Implementa nova estrutura de roteamento para o insert de aferição (PL-3623)
 * Adiciona endpoint de serviço de pneus na API (PL-3645)
 * Cria endpoint para listagem de pneus na API (PL-3646)
+
 
 ### Refactors
 * Melhora busca de indicadores consolidados e do extrato dos mapas (PL-3613)